refactor: Organize screens into proper tab navigation structure
