+++ app/(tabs)/explore.tsx
